fixxx

+++ src/GMod_LeapFinger.cpp
@@ -124,7 +124,7 @@ int LeapFinger::Bone( lua_State *state ) {
 	Finger *finger = Get( state );
 	if ( !finger ) return 0;
 
-	LeapBone::Push( state, new Leap::Bone( finger->bone( ( Bone::Type ) LUA->CheckNumber() ) ) );
+	LeapBone::Push( state, new Leap::Bone( finger->bone( ( Bone::Type ) ( ( int ) LUA->CheckNumber() ) ) ) );
 
 	return 1;
 }
